+++ sweetpad-vscode/CHANGELOG.md
@@ -5,7 +5,7 @@ New features, improvements and bug fixes for SweetPad are documented in this fil
 ## [0.2.15] - 2026-08-26
 
 - Fix missing schemes from Swift packages a project references ([#327](https://github.com/sweetpad-dev/sweetpad/issues/327), thanks [@rssole](https://github.com/rssole))
-- Stop listing schemes Xcode doesn't offer for product-less packages
+- Fix wrong scheme names for a standalone Swift package
 
 ## [0.2.14] - 2026-08-26